# proto/gen/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
    /api/v1/ai/chat/conversation:
        get:
            tags:
                - AIService
            description: GetChatConversation returns the caller's AI chat conversation.
            operationId: AIService_GetChatConversation
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ChatConversation'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/ai/chat:sendMessage:
        post:
            tags:
                - AIService
            description: |-
                SendChatMessage sends a user message in the caller's AI chat conversation
                 and returns the stored user message together with the assistant's reply.
            operationId: AIService_SendChatMessage
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/SendChatMessageRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SendChatMessageResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/ai:transcribe:
        post:
            tags:
                - AIService
            description: Transcribe transcribes an audio file using an instance AI provider.
            operationId: AIService_Transcribe
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/TranscribeRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/TranscribeResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
...
components:
    schemas:
...
        ChatCitation:
            type: object
            properties:
                memo:
                    type: string
                    description: |-
                        The resource name of the cited memo.
                         Format: memos/{memo}
                snippet:
                    type: string
                    description: The quoted source text from the memo.
            description: ChatCitation references the memo a chat answer was grounded in.
        ChatConversation:
            type: object
            properties:
                messages:
                    type: array
                    items:
                        $ref: '#/components/schemas/ChatMessage'
                    description: The messages in the conversation, in chronological order.
                generationAvailable:
                    type: boolean
                    description: |-
                        Whether the instance has a generation model configured.
                         When false, the assistant cannot generate new replies.
            description: ChatConversation is the caller's single conversation thread with the AI assistant.
        ChatMessage:
            type: object
            properties:
                role:
                    enum:
                        - ROLE_UNSPECIFIED
                        - USER
                        - ASSISTANT
                    type: string
                    description: The author of the message.
                    format: enum
                content:
                    type: string
                    description: The text content of the message.
                createTime:
                    type: string
                    description: The time the message was created.
                    format: date-time
                citations:
                    type: array
                    items:
                        $ref: '#/components/schemas/ChatCitation'
                    description: Source citations backing the message. Only set on assistant messages.
            description: ChatMessage is a single message in an AI chat conversation.
...
        GoogleProtobufAny:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
...
        SendChatMessageRequest:
            required:
                - content
            type: object
            properties:
                content:
                    type: string
                    description: Required. The text content of the user's message.
        SendChatMessageResponse:
            type: object
            properties:
                userMessage:
                    allOf:
                        - $ref: '#/components/schemas/ChatMessage'
                    description: The stored user message.
                assistantMessage:
                    allOf:
                        - $ref: '#/components/schemas/ChatMessage'
                    description: The assistant's reply message.
...
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
...
        TranscribeRequest:
            required:
                - audio
            type: object
            properties:
                audio:
                    allOf:
                        - $ref: '#/components/schemas/TranscriptionAudio'
                    description: Required. Audio input.
        TranscribeResponse:
            type: object
            properties:
                text:
                    type: string
                    description: The transcribed text.
        TranscriptionAudio:
            type: object
            properties:
                content:
                    writeOnly: true
                    type: string
                    description: Inline audio bytes.
                    format: bytes
                uri:
                    type: string
                    description: URI for audio content. Reserved for future use.
                filename:
                    type: string
                    description: Optional. The uploaded filename.
                contentType:
                    type: string
                    description: Optional. The MIME type of the input audio.
...